+++ CHANGELOG.md
@@ -31,6 +31,9 @@ CHANGELOG
 - [sdk/python] Define `__all__` in modules for better IDE autocomplete.
   [#6351](https://github.com/pulumi/pulumi/pull/6351)
 
+- [automation/python] Fix a bug in nested configuration parsing.
+  [#6349](https://github.com/pulumi/pulumi/pull/6349)
+
 ## 2.20.0 (2021-02-03)
 
 - [sdk/python] Fix `Output.from_input` to unwrap nested output values in input types (args classes), which addresses